Add inputrc and some other stuff
Add inputrc file, so that tab-completing a symlinked directory actually
behaves consistently with tab-completing a regular directory.

There are also apparently a whole bunch of other changes that I made. I
forgot what I was trying to do though, so I'll leave that to future self
to sort through (sorry).